requirements.txt
django==1.8.4
djangorestframework==3.2.3
markdown==2.6.2
django-filter==0.11.0
django-colorful==1.1.0
mysqlclient==1.3.6
djangorestframework-xml==1.2.0
pillow==2.9.0
